@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Http, Response, Headers, RequestOptions } from '@angular/http';
+// import { Http, Response, Headers, RequestOptions } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import '../rxjs-extensions';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable()
 export class CountryService {
@@ -9,7 +10,7 @@ export class CountryService {
   private readonly COUNTRY_SERVICE_URL = '/country';
   // private readonly COUNTRY_SERVICE_URL = 'http://services.groupkt.com/country';
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
   getAllCountries(): Observable<any> {
     // example header (not necessary)
@@ -47,18 +48,17 @@ export class CountryService {
     return observable;
   }
 
-  private extractData(res: Response) {
-    const body = res.json();
+  private extractData(res: any) {
     // if the result field is absent or does not have a value
-    if (body.RestResponse.result == null) {
-      throw(body.RestResponse.messages && body.RestResponse.messages.length > 1 && body.RestResponse.messages[1]
+    if (res.RestResponse.result == null) {
+      throw(res.RestResponse.messages && res.RestResponse.messages.length > 1 && res.RestResponse.messages[1]
         || 'Unknown error');
     }
 
-    return body.RestResponse.result;
+    return res.RestResponse.result;
   }
 
-  private handleError(error: Response | any) {
+  private handleError(error: any) {
     // In a real world app, we might use a remote logging infrastructure
     let errMsg: string;
     if (error instanceof Response) {
@@ -67,7 +67,7 @@ export class CountryService {
       errMsg = error.message ? error.message : error.toString();
     }
     // console.error(errMsg);
-    return Observable.throw(errMsg);
+    return Observable.throwError(errMsg);
   }
 
 }
